pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 79
target-version = ['py37']

[tool.poetry]
name = "grand-challenge.org"
version = "0.1.0"
description = ""
authors = ["James Meakin <code@jmsmkn.com>"]

[tool.poetry.dependencies]
python = ">=3.7"
"beautifulsoup4" = "*"
celery = "*"
redis = "*"
django = "<2.3"
django-countries = "*"
django-crispy-forms = "*"
django-userena-ce = "*"
djangorestframework = "*"
docker = "*"
matplotlib = "*"
"oauth2" = "*"
python-magic = "*"
python-memcached = "*"
pytz = "*"
social-auth-app-django = "*"
gunicorn = "*"
django-celery-email = "*"
nbconvert = "*"
simpleitk = "*"
django-celery-beat = "*"
django-favicon-plus = "*"
"psycopg2" = "*"
"django-select2" = "*"
django-celery-results = "*"
django-summernote = "*"
bleach = "*"
jsonschema = "*"
tldextract = "*"
tifffile = "==2019.1.4"
sorl-thumbnail = "*"
django-autocomplete-light = "*"
django-storages = "*"
boto3 = "*"
whitenoise = "*"
brotli = "*"
djangorestframework-guardian = "*"
django-extensions = "*"
django-simple-history = "*"
sentry-sdk = "*"
django-cors-headers = "*"
pyvips = "*"
django-speedinfo = "*"

[tool.poetry.dev-dependencies]
pytest-django = "*"
pytest-cov = "*"
pytest-mock = "*"
factory-boy = "*"
django-debug-toolbar = "*"
black = "==19.10b0"
sphinx-autobuild = "*"
sphinx = "*"
pyupgrade = "*"
pytest-xdist = "*"
sphinx-autodoc-typehints = "*"
werkzeug = "*"
sphinx-rtd-theme = "*"
flake8 = "^3.7"
flake8-bugbear = "^19.8"
flake8-import-order = "^0.18"
pep8-naming = "^0.8"
flake8-docstrings = "^1.5"
mccabe = "^0.6"
